Moved inspectors to directory
