+++ merlin/inspectors/circuitCounter.cc
