+++ merlin/inspectors/circuitCounter.h
